back: change pass endpoint

+++ app/src/main/WebInterfaceHandler.java
@@ -1,12 +1,15 @@
 package main;
 
 import com.google.gson.*;
+import com.google.gson.stream.JsonReader;
 import com.sun.net.httpserver.*;
 
 import java.io.*;
 import java.nio.charset.StandardCharsets;
 import java.util.Arrays;
 
+import entites.structures.ChangePassword;
+import services.Authentication;
 import services.CustomLogger;
 import entites.*;
 import services.DBConnect;
@@ -17,9 +20,8 @@ public class WebInterfaceHandler implements HttpHandler {
     public void handle(HttpExchange httpExchange) throws IOException {
         logger = new CustomLogger(WebInterfaceHandler.class.getName());
 
-        OutputStream responseStream = httpExchange.getResponseBody();
-        Headers responseHeaders = httpExchange.getResponseHeaders();
-
+        InputStream requestStream = httpExchange.getRequestBody();
+        JsonReader requestReader = new JsonReader(new InputStreamReader(requestStream));
         String method = httpExchange.getRequestMethod();
         String path = httpExchange.getRequestURI().getPath();
         String remoteAddr = httpExchange.getRequestHeaders()
@@ -29,19 +31,44 @@ public class WebInterfaceHandler implements HttpHandler {
 
         String params = httpExchange.getRequestURI().getQuery();
 
+        OutputStream responseStream = httpExchange.getResponseBody();
+        Headers responseHeaders = httpExchange.getResponseHeaders();
         Gson responseBody = new GsonBuilder().setPrettyPrinting().create();
+        int responseCode = 200;
         StringBuilder responseString = new StringBuilder();
+
+        User user;
 
         switch (path) {
             case "/api/users":
-                User user = DBConnect.getUser(httpExchange.getPrincipal().getUsername());
+                user = DBConnect.getUser(httpExchange.getPrincipal().getUsername());
                 responseString.append(responseBody.toJson(user));
                 break;
             case "/api/users/logout":
-                String cookies = httpExchange.getRequestHeaders().get("Cookie").get(0);
-                String token = cookies.split("=")[1];
-                DBConnect.deleteSession(token);
+                if (method.equals("DELETE")) {
+                    String cookies = httpExchange.getRequestHeaders().get("Cookie").get(0);
+                    String token = cookies.split("=")[1];
+                    DBConnect.deleteSession(token);
+                }
                 break;
+            case "/api/users/changeInfo":
+                if (method.equals("PATCH")) {
+                    user = DBConnect.getUser(httpExchange.getPrincipal().getUsername());
+                    try {
+                        ChangePassword passwordData = new Gson().fromJson(requestReader, ChangePassword.class);
+                        if (Authentication.checkPass(passwordData.oldPassword.toCharArray(), DBConnect.getUsersPassword(user.getUsername()))) {
+                            user.setPassword(Authentication.hashPass(passwordData.newPassword.toCharArray(), Authentication.createSalt()));
+                            DBConnect.updateUser(user);
+                            logger.info("Password for the user " + user.getUsername() + " was changed.");
+                        } else {
+                            responseCode = 401;
+                        }
+                    } catch (Exception e) {
+                        logger.warning(e.getMessage());
+                    }
+                }
+                break;
+                // TODO Create user
             case "/api/tickets":
                 String[] paramsArr = params.split("&");
                 logger.info(Arrays.toString(paramsArr));
@@ -51,7 +78,7 @@ public class WebInterfaceHandler implements HttpHandler {
         }
 
         responseHeaders.add("Content-Type", "application/json");
-        httpExchange.sendResponseHeaders(200, responseString.length());
+        httpExchange.sendResponseHeaders(responseCode, responseString.length());
         responseStream.write(responseString.toString().getBytes(StandardCharsets.UTF_8));
         responseStream.flush();
         httpExchange.close();
